enzyme working
