+++ babel.config.json
@@ -1,0 +1,4 @@
+{
+  "presets": ["@babel/preset-react", "@babel/preset-env"],
+  "plugins": ["@babel/plugin-transform-react-jsx"]
+}
