+++ client/css/styleMock.js
@@ -1,0 +1,2 @@
+// So JEST doesn't render CSS as export
+module.exports = {};
