+++ jest.config.js
@@ -1,0 +1,5 @@
+module.exports = {
+  moduleNameMapper: {
+    '\\.(css|less)$': '<rootDir>/client/css/styleMock.js',
+  }
+};
