+++ webpack.config.js
@@ -24,11 +24,7 @@ module.exports = {
         include : src,
         exclude: /node_modules/,
         use: {
-          loader: 'babel-loader',
-          options: {
-            presets: ['@babel/preset-env', '@babel/preset-react'],
-            plugins: ['@babel/plugin-proposal-object-rest-spread']
-          }
+          loader: 'babel-loader'
         }
       }
     ]
